requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.12.0
django-restframework==0.0.1
djangorestframework==3.13.1
idna==3.3
pkg_resources==0.0.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
